1.[Bug Fixed] #123094 home page clickable area 2.[Bug Fixed] #123228 Android APP sometimes will crash(1/20) 3.[Bug Fixed] #124115 Add MESH page, the MPro Mesh extender doesn't have mode switch button

+++ app/src/main/java/zyxel/com/multyproneo/model/WiFiSettingInfo.kt
@@ -39,14 +39,14 @@ data class Radio
 
 data class Stats
 (
-    val BytesReceived: Int = 0,
-    val BytesSent: Int = 0,
-    val DiscardPacketsReceived: Int = 0,
-    val DiscardPacketsSent: Int = 0,
-    val ErrorsReceived: Int = 0,
-    val ErrorsSent: Int = 0,
-    val PacketsReceived: Int = 0,
-    val PacketsSent: Int = 0,
+    val BytesReceived: Long = 0,
+    val BytesSent: Long = 0,
+    val DiscardPacketsReceived: Long = 0,
+    val DiscardPacketsSent: Long = 0,
+    val ErrorsReceived: Long = 0,
+    val ErrorsSent: Long = 0,
+    val PacketsReceived: Long = 0,
+    val PacketsSent: Long = 0,
     val X_ZYXEL_Rate: String = "N/A"
 )
